couple more bulk sim updates
