+++ ui/scss/core/components/_gear_picker.scss
@@ -172,14 +172,6 @@
 		.selector-modal-tabs {
 			border-bottom: 0;
 		}
-
-		.btn-danger {
-			margin: calc(-0.5 * var(--bs-modal-header-padding-y)) var(--spacer-3) calc(-0.5 * var(--bs-modal-header-padding-y)) auto;
-
-			& + .btn-close {
-				margin-left: 0;
-			}
-		}
 	}
 }
 
@@ -260,7 +252,7 @@
 
 .selector-modal-list {
 	width: 100%;
-	max-height: 65vh;
+	height: 65vh;
 	overflow-y: scroll;
 	overflow-x: hidden;
 	padding: 0;
