+++ ui/scss/core/individual_sim_ui/bulk/_bulk_item_picker.scss
@@ -1,0 +1,18 @@
+.bulk-item-picker.item-picker-root {
+	padding: var(--spacer-2);
+	border: var(--border-default);
+
+	.item-picker-labels-container {
+		padding-top: 0;
+	}
+
+	.remove-batch-item-btn {
+		@extend .btn;
+		@extend .btn-sm;
+		@extend .btn-reset;
+
+		margin-right: calc(-1 * var(--spacer-2));
+		margin-top: calc(-1 * var(--spacer-2));
+		align-self: flex-start;
+	}
+}
