+++ ui/scss/core/individual_sim_ui/bulk/index.scss
@@ -1,0 +1,1 @@
+@import './bulk_item_picker';
