+++ ui/scss/core/individual_sim_ui/index.scss
@@ -1,35 +1,37 @@
 // Import Sim UI and system-wide styles
-@import "../sim_ui/index";
+@import '../sim_ui/index';
 
 // Import Individual Sim UI Styles
-@import "./shared";
+@import './shared';
+@import './bulk';
 
 // Import component styles
-@import "../components/boolean_picker";
-@import "../components/character_stats";
-@import "../components/detailed_results";
-@import "../components/encounter_picker";
-@import "../components/exporters";
-@import "../components/enum_picker";
-@import "../components/bulk";
-@import "../components/gear_picker";
-@import "../components/icon_enum_picker";
-@import "../components/icon_picker";
-@import "../components/importers";
-@import "../components/item_swap_picker";
-@import "../components/multi_icon_picker";
-@import "../components/number_list_picker";
-@import "../components/number_picker";
-@import "../components/raid_sim_action";
-@import "../components/raid_target_picker";
-@import "../components/saved_data_manager";
-@import "../components/settings_menu";
-@import "../components/stat_weights_action";
-@import "../components/unit_picker";
+@import '../components/boolean_picker';
+@import '../components/character_stats';
+@import '../components/detailed_results';
+@import '../components/encounter_picker';
+@import '../components/exporters';
+@import '../components/enum_picker';
+@import '../components/bulk';
+@import '../components/gear_picker';
+@import '../components/icon_enum_picker';
+@import '../components/icon_picker';
+@import '../components/importers';
+@import '../components/item_swap_picker';
+@import '../components/multi_icon_picker';
+@import '../components/number_list_picker';
+@import '../components/number_picker';
+@import '../components/raid_sim_action';
+@import '../components/raid_target_picker';
+@import '../components/saved_data_manager';
+@import '../components/settings_menu';
+@import '../components/stat_weights_action';
+@import '../components/unit_picker';
 
-@import "../components/individual_sim_ui/settings_tab";
-@import "../components/individual_sim_ui/rotation_tab";
-@import "../components/individual_sim_ui/talents_tab";
+@import '../components/individual_sim_ui/bulk_tab';
+@import '../components/individual_sim_ui/settings_tab';
+@import '../components/individual_sim_ui/rotation_tab';
+@import '../components/individual_sim_ui/talents_tab';
 
-@import "../talents/hunter_pet";
-@import "../talents/talents_picker";
+@import '../talents/hunter_pet';
+@import '../talents/talents_picker';
