+++ ui/scss/index.scss
@@ -43,7 +43,6 @@
 
 // Custom Bootstrap style overrides
 @import 'shared/bootstrap_style_overrides';
-@import 'shared/bootstrap_style_overrides';
 
 // Import other plugin styles
 @import 'tippy.js/dist/tippy';
